requirements.txt
adbutils==0.11.0
aiohttp==3.8.3
aiosignal==1.3.1
apkutils2==1.0.0
async-timeout==4.0.2
attrs==22.2.0
av==8.1.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==2.1.1
cigam==0.0.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
deprecation==2.1.0
efficientnet-pytorch==0.7.1
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
huggingface-hub==0.11.1
idna==3.4
imageio==2.24.0
kiwisolver==1.4.4
lightning-utilities==0.5.0
matplotlib==3.6.3
multidict==6.0.4
munch==2.5.0
networkx==3.0
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
pretrainedmodels==0.7.4
py==1.11.0
pyelftools==0.29
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.9.0
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
retry==0.9.2
scikit-image==0.19.3
scipy==1.10.0
scrcpy==0.0.4
segmentation-models-pytorch==0.3.2
six==1.16.0
tifffile==2022.10.10
timm==0.6.12
torch==1.13.1
torchmetrics==0.11.0
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
whichcraft==0.6.1
xmltodict==0.13.0
yarl==1.8.2
